Rename analysis module
