+++ OTAnalytics/plugin_ui/main_application.py
@@ -1,7 +1,7 @@
 from OTAnalytics.adapter_intersect.intersect import (
     ShapelyIntersectImplementationAdapter,
 )
-from OTAnalytics.application.analysis.analysis import (
+from OTAnalytics.application.analysis.intersect import (
     RunIntersect,
     RunSceneEventDetection,
 )
